requirements.txt
python-dotenv
streamlit
replicate
google.generativeai
